Forgot password, Reset password and email sending added

+++ auth-service/src/main/java/com/hajela/authservice/exceptions/UserIdNotFoundException.java
@@ -2,7 +2,7 @@ package com.hajela.authservice.exceptions;
 
 public class UserIdNotFoundException extends RuntimeException {
 
-    public UserIdNotFoundException(Integer userId) {
+    public UserIdNotFoundException(Long userId) {
         super("User not found with Id: " + userId);
     }
 }
